Adding a new splash page
	The Timeclock class did not have a way to query for the current
status of all clockedin users. The new `statusAll` member function takes
on that role returning an array with the values for `clockedin`,
`first_name`, `last_name`, and `purpose`. These are used by the new
index.php file which should display the running log of clockedin users
and provide a button to log into the site.

	Additionally this commit changes a lot of whitespace in the
files touched. Apologies for the messy patch.